requirements.txt
websocket-client
requests
playwright
streamlit
pandas
plotly
numpy
statsmodels